src/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
PyYAML==6.0.1
Markdown==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
click==8.1.7
itsdangerous==2.1.2
requests==2.31.0
feedparser==6.0.10
python-dateutil==2.8.2
Pygments==2.15.1